+++ docease/src/main/java/com/example/docease/repositories/ChatMessageRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessage, Integer> {
     List<ChatMessage> findBySender(User sender);
     List<ChatMessage> findByReceiver(User receiver);
