+++ docease/src/main/java/com/example/docease/repositories/DoctorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DoctorRepository extends JpaRepository <Doctor, Long>{
+public interface DoctorRepository extends JpaRepository <Doctor, Integer>{
 
     List<Doctor> findBySpecialization(String specialization);
