+++ docease/src/main/java/com/example/docease/repositories/PatientRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PatientRepository extends JpaRepository <Patient, Long>{
-    Optional<Patient> findByUserId(Long userId);
+public interface PatientRepository extends JpaRepository <Patient, Integer>{
+    Optional<Patient> findByUserId(Integer userId);
 }
